addressed more code review comments
