+++ src/app/AppRegistry.tsx
@@ -24,7 +24,7 @@ import { GalleriesForYouScreen } from "app/Scenes/GalleriesForYou/GalleriesForYo
 import { HomeViewScreen, homeViewScreenQuery } from "app/Scenes/HomeView/HomeView"
 import { HomeViewSectionScreenQueryRenderer } from "app/Scenes/HomeViewSectionScreen/HomeViewSectionScreen"
 import { ConversationQueryRenderer } from "app/Scenes/Inbox/Screens/Conversation"
-import { InfiniteDiscovery } from "app/Scenes/InfiniteDiscovery/InfiniteDiscovery"
+import { InfiniteDiscoveryView } from "app/Scenes/InfiniteDiscovery/InfiniteDiscovery"
 import { AddMyCollectionArtist } from "app/Scenes/MyCollection/Screens/Artist/AddMyCollectionArtist"
 import { MyCollectionArtworkEditQueryRenderer } from "app/Scenes/MyCollection/Screens/ArtworkForm/Screens/MyCollectionArtworkEdit"
 import { MyCollectionCollectedArtistsPrivacyQueryRenderer } from "app/Scenes/MyCollection/Screens/CollectedArtistsPrivacy/MyCollectionCollectedArtistsPrivacy"
@@ -580,13 +580,11 @@ export const modules = defineModules({
     Queries: [InboxScreenQuery],
   }),
   InfiniteDiscovery: reactModule({
-    Component: InfiniteDiscovery,
+    Component: InfiniteDiscoveryView,
     options: {
       fullBleed: true,
+      hidesBackButton: true,
       hidesBottomTabs: true,
-      screenOptions: {
-        headerShown: false,
-      },
     },
   }),
   Inquiry: reactModule({
